requirements.txt
python-dotenv>=1.0.0
pandas>=1.3.1
pyyaml>=5.4.0
openpyxl>=3.1.2
Pyarrow>=15.0.0
sqlalchemy>=1.4.20
alembic>=1.13.1
psycopg2-binary
